Bug fix: Ticket#246 Large FIXED Constants do not compile

+++ openpearl-code/compiler/src/org/smallpearl/compiler/ConstantExpressionEvaluatorVisitor.java
@@ -143,7 +143,7 @@ public  class ConstantExpressionEvaluatorVisitor extends SmallPearlBaseVisitor<V
         }
 
         if ( ctx.fixedConstant() != null) {
-            int curval = sign * Integer.parseInt(ctx.fixedConstant().IntegerConstant().toString());
+            long curval = sign * Long.parseLong(ctx.fixedConstant().IntegerConstant().toString());
             int curlen =   m_currentSymbolTable.lookupDefaultFixedLength();
 
             if ( ctx.fixedConstant().fixedNumberPrecision() != null ) {
